+++ packages/peer-discovery-mdns/src/query.ts
@@ -1,6 +1,7 @@
 import { logger } from '@libp2p/logger'
 import { peerIdFromString } from '@libp2p/peer-id'
-import { multiaddr, type Multiaddr } from '@multiformats/multiaddr'
+import { isPrivate } from '@libp2p/utils/multiaddr/is-private'
+import { multiaddr, type Multiaddr, protocols } from '@multiformats/multiaddr'
 import type { PeerInfo } from '@libp2p/interface/peer-info'
 import type { Answer, StringAnswer, TxtAnswer } from 'dns-packet'
 import type { MulticastDNS, QueryPacket, ResponsePacket } from 'multicast-dns'
@@ -9,7 +10,7 @@ const log = logger('libp2p:mdns:query')
 
 export function queryLAN (mdns: MulticastDNS, serviceTag: string, interval: number): ReturnType<typeof setInterval> {
   const query = (): void => {
-    log('query', serviceTag)
+    log.trace('query', serviceTag)
 
     mdns.query({
       questions: [{
@@ -40,6 +41,16 @@ export function gotResponse (rsp: ResponsePacket, localPeerName: string, service
     }
   })
 
+  // according to the spec, peer details should be in the additional records,
+  // not the answers though it seems go-libp2p at least ignores this?
+  // https://github.com/libp2p/specs/blob/master/discovery/mdns.md#response
+  rsp.additionals.forEach((answer) => {
+    switch (answer.type) {
+      case 'TXT': txtAnswers.push(answer); break
+      default: break
+    }
+  })
+
   if (answerPTR == null ||
     answerPTR?.name !== serviceTag ||
     txtAnswers.length === 0 ||
@@ -63,7 +74,7 @@ export function gotResponse (rsp: ResponsePacket, localPeerName: string, service
 
     return {
       id: peerIdFromString(peerId),
-      multiaddrs,
+      multiaddrs: multiaddrs.map(addr => addr.decapsulateCode(protocols('p2p').code)),
       protocols: []
     }
   } catch (e) {
@@ -92,20 +103,45 @@ export function gotQuery (qry: QueryPacket, mdns: MulticastDNS, peerName: string
       data: peerName + '.' + serviceTag
     })
 
-    multiaddrs.forEach((addr) => {
-      // spec mandates multiaddr contains peer id
-      if (addr.getPeerId() != null) {
+    multiaddrs
+      // mDNS requires link-local addresses only
+      // https://github.com/libp2p/specs/blob/master/discovery/mdns.md#issues
+      .filter(isLinkLocal)
+      .forEach((addr) => {
+        const data = 'dnsaddr=' + addr.toString()
+
+        // TXT record fields have a max data length of 255 bytes
+        // see 6.1 - https://www.ietf.org/rfc/rfc6763.txt
+        if (data.length > 255) {
+          log('multiaddr %a is too long to use in mDNS query response', addr)
+          return
+        }
+
+        // spec mandates multiaddr contains peer id
+        if (addr.getPeerId() == null) {
+          log('multiaddr %a did not have a peer ID so cannot be used in mDNS query response', addr)
+          return
+        }
+
         answers.push({
           name: peerName + '.' + serviceTag,
           type: 'TXT',
           class: 'IN',
           ttl: 120,
-          data: 'dnsaddr=' + addr.toString()
+          data
         })
-      }
-    })
+      })
 
-    log('responding to query')
+    log.trace('responding to query')
     mdns.respond(answers)
   }
+}
+
+function isLinkLocal (ma: Multiaddr): boolean {
+  // match private ip4/ip6 & loopback addresses
+  if (isPrivate(ma)) {
+    return true
+  }
+
+  return false
 }